// src/types/stream.types.ts
export interface Stream {
  user_login: string;
  github_url: string;
  twitter_url: string;
  instagram_url: string;
  linkedin_url: string;
  discord_url: string;
  id: string;
  user_id: string;
  user_name: string;
  title: string;
  viewer_count: number;
  started_at: string;
  thumbnail_url: string;
  profile_image_url: string;
  description: string;
  stream_id: string;
  duration: string;
  tags?: string[];
}